event moving npc but sinchronously

+++ WanagiYata/Assets/Scripts/Model/EventModel.cs
@@ -8,8 +8,8 @@ namespace Assets.Scripts.Model
     public class EventModel
     {
         public string Target { get; set; }
-        public int? DestinationX { get; set; }
-        public int? DestinationY { get; set; }
+        public float? DestinationX { get; set; }
+        public float? DestinationY { get; set; }
         public string Interaction { get; set; }
         public bool WaitPrevious { get; set; }
     }
